+++ frontend/src/components/InputForm.vue
@@ -13,7 +13,7 @@
       <v-row class="d-flex justify-center w-100">
         <v-col sm="10" md="8" class="w-100">
           <v-row class="mb-4 d-flex justify-center w-100">
-            <v-col cols="10" class="w-100">
+            <v-col cols="12" class="w-100">
               <ReportStepper />
             </v-col>
           </v-row>
@@ -365,7 +365,7 @@ import VueDatePicker from '@vuepic/vue-datepicker';
 import '@vuepic/vue-datepicker/dist/main.css';
 import PrimaryButton from './util/PrimaryButton.vue';
 import Spinner from './Spinner.vue';
-import ReportStepper from './util/ReportStepper.vue';
+import ReportStepper from './util/ReportStepper/Stepper.vue';
 import ApiService, { ISubmission } from '../common/apiService';
 import { useCodeStore } from '../store/modules/codeStore';
 import { authStore } from '../store/modules/auth';
