--- frontend/src/components/util/ReportStep.vue
@@ -1,74 +1,0 @@
-<template>
-  <v-col class="d-flex d-flex-col justify-center align-center">
-    <div
-      class="circle"
-      :class="{
-        active: value == stage,
-        available: value != stage,
-        disabled: disabled,
-      }"
-      @click="handleClick"
-      :data-testid="'report-step-' + value"
-      :disabled="disabled"
-    >
-      {{ index + 1 }}
-    </div>
-  </v-col>
-  <v-col v-if="index !== 2" class="d-flex justify-center align-center">
-    <div class="dash disabled"></div>
-  </v-col>
-</template>
-
-<script setup lang="ts">
-import { storeToRefs } from 'pinia';
-import { useReportStepperStore } from '../../store/modules/reportStepper';
-import { useRouter } from 'vue-router';
-
-const router = useRouter();
-const { stage } = storeToRefs(useReportStepperStore());
-const props = defineProps(['value', 'label', 'url', 'index', 'disabled']);
-
-const handleClick = () => {
-  if (props.disabled) return;
-  router.push(props.url);
-};
-</script>
-
-<style scoped>
-.circle {
-  height: 60px;
-  width: 60px;
-  color: #ffffff;
-  border-radius: 50%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  &.available {
-    background-color: #00336633;
-    cursor: pointer;
-  }
-
-  &.active {
-    background-color: #003366;
-    cursor: pointer;
-  }
-
-  &.disabled {
-    background-color: #aaaaaa;
-    cursor: not-allowed;
-  }
-}
-
-.dash {
-  height: 1px;
-  width: 100%;
-  background-color: #003366;
-  padding-left: 5px;
-  padding-right: 5px;
-
-  &.disabled {
-    background-color: #aaaaaa;
-  }
-}
-</style>
